refactor: 비대해진 ArchivedPageInfo를 분리한다
 - 내부에 정의된 ArchivedSnapshots, ClosestSnapshot을 별도의 클래스로
   분리하고, 접근 제어자를 가능한 좁힌다
 - 테스트용으로만 쓰는 ArchivedPageInfoBuilder를 테스트 패키지로 옮긴다
