+++ src/main/java/org/gsh/genidxpage/service/dto/ArchivedPageInfo.java
@@ -38,96 +38,4 @@ public class ArchivedPageInfo {
     public String accessibleUrl() {
         return archivedSnapshots.getSnapshot().getUrl();
     }
-
-    public static class ArchivedPageInfoBuilder {
-        private String url;
-        private ArchivedSnapshots archivedSnapshots;
-        private String timestamp;
-
-        public static ArchivedPageInfoBuilder builder() {
-            return new ArchivedPageInfoBuilder();
-        }
-
-        public ArchivedPageInfoBuilder url(String url) {
-            this.url = url;
-            return this;
-        }
-
-        public ArchivedPageInfoBuilder withEmptyArchivedSnapshots() {
-            this.archivedSnapshots = null;
-            return this;
-        }
-
-        public ArchivedPageInfoBuilder withAccessibleArchivedSnapshots() {
-            this.archivedSnapshots = new ArchivedSnapshots(
-                new ClosestSnapshot("200",
-                    true,
-                    "http://localhost:8080/web/20230614220926/archives/2021/03",
-                    "20230614220926")
-            );
-            return this;
-        }
-
-        public ArchivedPageInfoBuilder timestamp(String timestamp) {
-            this.timestamp = timestamp;
-            return this;
-        }
-
-        public ArchivedPageInfo build() {
-            return new ArchivedPageInfo(url, archivedSnapshots, timestamp);
-        }
-    }
-
-    private static class ArchivedSnapshots {
-
-        @JsonProperty("closest")
-        private ClosestSnapshot snapshot;
-
-        public ArchivedSnapshots() {
-        }
-
-        public ArchivedSnapshots(ClosestSnapshot snapshot) {
-            this.snapshot = snapshot;
-        }
-
-        public ClosestSnapshot getSnapshot() {
-            return snapshot;
-        }
-    }
-
-    private static class ClosestSnapshot {
-
-        private String status;
-        private Boolean available;
-        @JsonProperty("url")
-        private String url;
-        @JsonProperty("timestamp")
-        private String timestamp;
-
-        public ClosestSnapshot() {
-        }
-
-        public ClosestSnapshot(String status, Boolean available, String url, String timestamp) {
-            this.status = status;
-            this.available = available;
-            this.url = url;
-            this.timestamp = timestamp;
-        }
-
-        public String getStatus() {
-            return status;
-        }
-
-        public Boolean getAvailable() {
-            return available;
-        }
-
-        public String getUrl() {
-            return url;
-        }
-
-        public String getTimestamp() {
-            return timestamp;
-        }
-    }
 }
